// app/layout/models/sales-restriction.model.ts
export class SalesRestriction {
    storeLocationSalesRestrictionID: number;
    storeLocationID: number;
    salesRestrictionDescription: string;
    salesRestrictFlag: boolean;
    prohibitDiscountFlag: boolean;
    minimumCustomerAge: number;
    minimumClerkAge: number;
    transactionLimit: number;
    posSalesRestrictID: number;
}